remove duplicate old tests and a few test fixes
